Add additional configuration options and document all options (#6)
* Add possibility to set a commit message

* Allow for disabling force push

If the action is pushing to the same branch, it is likely desirable to
do a regular push instead of a force which might overwrite other
changes in case the source has changed after the action triggered.

To not break the interface of the action, add a boolean flag to turn
on 'NO_FORCE_PUSH' mode.

* Fix output of branch the action will publish to

Likely referenced an old name of the current `PUBLISH_TO_BRANCH` env var,
and printed nothing.

* Document configuration options

+++ entrypoint.sh
@@ -26,7 +26,7 @@ marp ${MARP_ARGS}
 echo "✔  Built Successfully!"
 echo ""
 
-echo "   Publishing to ${GITHUB_REPOSITORY} ${REMOTE_BRANCH}..."
+echo "   Publishing to ${GITHUB_REPOSITORY} ${PUBLISH_TO_BRANCH}..."
 echo ""
 
 remote_repo="https://x-access-token:${GITHUB_TOKEN}@github.com/${GITHUB_REPOSITORY}.git" && \
@@ -36,8 +36,15 @@ git config user.email "marp-action@users.noreply.github.com" && \
 git add . && \
 git status && \
 curr_branch="$(git rev-parse --abbrev-ref HEAD)" && \
-git commit -m'action build' && \
-git push --force $remote_repo ${curr_branch}:${PUBLISH_TO_BRANCH}
+commit_msg="${COMMIT_MSG:-'Marp Action Build'}" && \
+git commit -m "${commit_msg}" 
+
+if [ "$NO_FORCE_PUSH" = true ]
+then 
+  git push $remote_repo ${curr_branch}:${PUBLISH_TO_BRANCH}
+else
+  git push --force $remote_repo ${curr_branch}:${PUBLISH_TO_BRANCH}
+fi
 
 echo "✔  Pushed Successfully!"
 echo ""
